docs(advanced): new section Routes
