+++ resources/views/pages/ru/resources/index.blade.php
@@ -7,7 +7,6 @@
             ['url' => '#create', 'label' => 'Создание'],
             ['url' => '#define', 'label' => 'Объявление'],
             ['url' => '#item', 'label' => 'Текущий элемент/модель'],
-            ['url' => '#routes', 'label' => 'Маршруты ресурса'],
             ['url' => '#modal', 'label' => 'Модальные окна'],
             ['url' => '#after', 'label' => 'Переход после сохранения'],
             ['url' => '#simple-pagination', 'label' => 'Simple pagination'],
@@ -181,61 +180,6 @@ class MoonShineServiceProvider extends ServiceProvider
 
 <x-moonshine::alert type="default" icon="heroicons.information-circle">
     Если элемента еще не существует (action create), то метод <code>getItem()</code> вернет <code>NULL</code>.
-</x-moonshine::alert>
-
-<x-sub-title id="routes">Маршруты ресурса</x-sub-title>
-
-<x-p>
-    В MoonShine у ресурса для различных действий зарегистрировано множество маршрутов:
-</x-p>
-
-<x-code language="php">
-$this->route('index'); // GET|HEAD - список записей
-$this->route('create'); // GET|HEAD - создание новой записи
-$this->route('store'); // POST - сохранение новой записи
-$this->route('edit', $resourceItem); // GET|HEAD - редактирование записи
-$this->route('update', $resourceItem); // PUT|PATCH - сохранение записи
-$this->route('destroy', $resourceItem); // DELETE - удаление записи
-$this->route('show', $resourceItem); // GET|HEAD - просмотр записи
-$this->route('query-tag', $queryTag); // GET|HEAD - список записей с применением быстрого фильтра / тега
-$this->route('update-column', $resourceItem); // PUT - сохранение поля записи
-</x-code>
-
-<x-p>
-    Через метод <code>resolveRoutes()</code> можно добавить или переопределять роуты по умолчанию.
-</x-p>
-
-<x-code language="php">
-namespace App\MoonShine\Resources;
-
-use App\Models\Post;
-use MoonShine\Resources\Resource;
-
-class PostResource extends Resource
-{
-    public static string $model = Post::class;
-
-    // ...
-
-    public function resolveRoutes(): void // [tl! focus:start]
-    {
-        parent::resolveRoutes();
-
-        Route::prefix('resource')->group(function (): void {
-            Route::get("{$this->uriKey()}/restore/{resourceItem}", function (Post $item) {
-                $item->restore();
-
-                return redirect()->back();
-            });
-        });
-    } // [tl! focus:end]
-
-    // ...
-}
-</x-code>
-
-<x-moonshine::alert type="default" icon="heroicons.book-open">
-    Для доступа к маршруту вне ресурса можно следующим способом <code>(new Resource())->route('index')</code>.
 </x-moonshine::alert>
 
 <x-sub-title id="modal">Модальные окна</x-sub-title>
